[#6] Move the extension and usecase to correct layer
