+++ settings.gradle.kts
@@ -9,4 +9,3 @@ pluginManagement {
 rootProject.name = "KMM-IC"
 include(":androidApp")
 include(":shared")
-//include(":jsonapi-kotlin:core")
